requirements.txt
langchain>=0.1.0,<0.2.0
langchain-core>=0.1.10,<0.2.0
langchain-openai>=0.0.2,<0.1.0
langgraph>=0.0.20,<0.1.0
pydantic>=2.5.2,<3.0.0
openai>=1.6.1,<2.0.0
ipython>=8.18.1,<9.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.25.1,<3.0.0

